requirements.txt
Flask
Flask-SocketIO
openai-whisper
googletrans
numpy==1.21.0
pydub
